+++ app/000_init.go
@@ -1,4 +1,4 @@
 package app
 
 // This file must be lexicographically the first one in the app module, so it's guaranteed that the following dependency is the first to be imported
-import _ "github.com/axelarnetwork/axelar-core/app/encoding_fix"
+import _ "github.com/axelarnetwork/axelar-core/app/encoding_override"
